third_party/WebKit/LayoutTests/http/tests/webfont/webfont-performance-duration-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS entry.startTime >= waitMillisBeforeFetching is true
Hello webfont
